Feature Branch: June Release -- GenRL Integration
* Revert "Revert "rework to use genrl backend ""

This reverts commit decff722d5e18e6655a0bc1245c13d2d568ebc08.

* Update account-kit deps for permission builder

* Avoid signer call on every request

* Read new Gossip shape

Several updates, the most important of which is reading the new gossip shape from the genrl dht.

* Update dht_pub.py to read new shape
* Update golang processor to write a dataset attribute
* Remove lots of unused code

* script that runs a dht forever

* Remove create signer for user

* Point at genrl contract

* Remove "April" branching logic

Step 0 of updating how we compute nodes connected; cleaning up old branching logic.

This change is about the minimum amount of work needed to remove "april release" branching logic.
We had this in place initially because we swapped the DB schema between March and April, but April is now the standard implementation, and the old databases are not accessed anymore.

APRIL_RELEASE is 1 in every service already.
https://github.com/gensyn-ai/infra/blob/34f237a19c9edc47be260f835fad396205b10e88/terraform/projects/swarm-ui/main.tf

Expected changes: none. This PR effectively removes unused code.

* Calculate nodes connected based on time buckets

Every time a reward is submitted, the nodes connected is incremented.

The math/math-hard implementation is to group those rewards by round+stage, and display the number of rewards submitted in the previous round/stage.

If we're in the "June release" environment, the nodes connected are calculated by the number of rewards submitted in an hour time frame rather than round/stage, which decouples that stat from round/stage.

To prevent the value from dropping too much, we query for the 2 latest entries, and return the least recent one.

* Cniessigma/leaner meaner bootnodes

* GenRL as RL-swarm submodule

* Submodule GenRL, make and copy config

only backup if config doesn't exist
copy code to match latest pr

* update submodule commit to latest

---------

Co-authored-by: Eddie Nuno <GOD11341258@gmail.com>
Co-authored-by: Tristin Forbus <tforbus@users.noreply.github.com>
Co-authored-by: Christopher Nies <cniessigma@gmail.com>
GitOrigin-RevId: f3ecb470a0b8b3f98b01fa0a9679138098d3c2cb